requirements.txt
astrocalc>=0.4.1
astropy>=4.2
black>=22.8
bokeh>=2.4.3
Django>=3.2
django-filter>=2.4.0
django-geoip2-extras>=2.0.2
django-stubs>=1.12.0
djangorestframework>=3.12.4
djangorestframework-stubs>=1.7.0
flake8>=5.0.4
fundamentals>=2.3.0
#geoip2>=4.5.0
julian>=0.14
Markdown>=3.2.2
matplotlib>=3.3.3
mod-wsgi-standalone>=4.7.1
multiprocess>=0.70.11.1
#mysql>=0.0.2
#mysql-connector-python>=8.0.22
mysqlclient>=2.0.3
pandas>=1.2.4
pycountry>=20.7.3
pylint>=2.15.0
pylint-django>=2.5.3
python-dotenv>=0.17.0
types-PyYAML>=6.0.11
types-requests>=2.28.9
